requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pytest==7.4.3
httpx==0.24.1
alembic==1.12.1